+++ lib/routes/login.js
@@ -7,6 +7,7 @@ var router = new express.Router();
 
 var config = require('../../config');
 var user_db = require('../models/user');
+var get_user_ip = require('../utils/get-user-ip');
 
 
 router.get('/login', function(req, res) {
@@ -15,12 +16,7 @@ router.get('/login', function(req, res) {
 
 router.post('/login', function(req, res) {
     res.type('text/plain');
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers['x-forwarded-for'];
-    if (ip_addr) {
-        var list = ip_addr.split(',');
-        user_ip = list[list.length - 1];
-    }
+    var user_ip = get_user_ip(req);
 
     var secure_text = crypto.createHash('sha1')
                       .update(req.body.digits + config.securetrip_salt)
