+++ lib/utils/check-ip-validity.js
@@ -3,6 +3,7 @@
 var fs = require('fs');
 
 var user_db = require('../models/user');
+var get_user_ip = require('./get-user-ip');
 
 /* check_ip_validity:
     - checks if ip is banned
@@ -12,12 +13,7 @@ var user_db = require('../models/user');
 */
 module.exports = function(req, callback) {
     /* get IP */
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers['x-forwarded-for'];
-    if (ip_addr && !user_ip) {
-        var list = ip_addr.split(',');
-        user_ip = list[list.length - 1];
-    }
+    var user_ip = get_user_ip(req);
 
     /* lookup IP */
     user_db
