+++ lib/utils/format-special.js
@@ -3,17 +3,13 @@
 var geoip = require('geoip');
 var City = geoip.City;
 var city = new City('GeoLiteCity.dat');
+var get_user_ip = require('./get-user-ip');
 
 var default_country = "UN";
 
 function get_country(req, data, callback) {
     /* get IP */
-    var user_ip = req.connection.remoteAddress;
-    var ip_addr = req.headers['x-forwarded-for'];
-    if (ip_addr) {
-        var list = ip_addr.split(',');
-        user_ip = list[list.length - 1];
-    }
+    var user_ip = get_user_ip(req);
     // geoip
     city.lookup(user_ip, function(err, c_data) {
 	    if (err) {
