+++ lib/utils/populate-post.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var fs = require('fs');
+var get_user_ip = require('./get-user-ip');
 
 /* populate_post:
     - converts a POST request into a data object with fields:
@@ -21,13 +22,7 @@ module.exports = function(req, data, callback) {
         data.body = req.body.body || "";
         data.name = req.body.name || "Anonymous";
         data.date = (new Date()).toString();
-        var user_ip = req.connection.remoteAddress;
-        var ip_addr = req.headers["x-forwarded-for"];
-        if (ip_addr) {
-            var list = ip_addr.split(",");
-            user_ip = list[list.length - 1];
-        }
-        data.ip = user_ip;
+        data.ip = get_user_ip(req);
 
         /* populate initial file details */
         if (req.files &&
